updates and fixes to code for build. Fixed type errors and component issues. Added new components so that pages were no longer having issues.
