+++ app/_components/usercomps/userfavs.tsx
@@ -1,4 +1,4 @@
-import {Grid} from "@mantine/core";
+import {Grid, GridCol} from "@mantine/core";
 import MarketCard from "@/app/_components/marketaccordian/marketcard";
 import {MarketGridProps, MarketsInterface} from "@/app/_types/interfaces";
 import React from "react";
@@ -9,9 +9,9 @@ export default function UserFavs({data} : MarketGridProps){
     return (
         <Grid gutter="xl">
             {data.map((market: MarketsInterface) => (
-                <Grid.Col span={{base: 12, sm: 6, md: 4}} key={market.id}>
+                <GridCol span={{base: 12, sm: 6, md: 4}} key={market.id}>
                     <MarketCard key={market.id} market={market}/>
-                </Grid.Col>
+                </GridCol>
             ))}
         </Grid>
     )
