+++ app/contact/page.tsx
@@ -39,7 +39,7 @@ export default function Page() {
   return (
     <>
       <Modal opened={opened} onClose={close} title="Message Sent!" centered>
-        <Text>Thank you for your message! We'll get back to you shortly.</Text>
+        <Text>Thank you for your message! We&apos;ll get back to you shortly.</Text>
       </Modal>
 
       <AppShell>
@@ -52,7 +52,7 @@ export default function Page() {
                 Contact Us
               </Title>
               <Text ta="center" mb="md" c="gray.7">
-                Have a question or want to join our platform? Reach out and we'll get back to you shortly.
+                Have a question or want to join our platform? Reach out and we&apos;ll get back to you shortly.
               </Text>
 
               <form
@@ -121,7 +121,7 @@ export default function Page() {
                     Are You a Farmers Market?
                   </Title>
                   <Text size="sm" c="gray.7">
-                    We'd love to feature your market in our directory. Use the form above to tell us more.
+                    We&apos;d love to feature your market in our directory. Use the form above to tell us more.
                   </Text>
                 </Card>
               </Grid.Col>
@@ -132,7 +132,7 @@ export default function Page() {
                     Are You a Vendor?
                   </Title>
                   <Text size="sm" c="gray.7">
-                    Want to showcase your products on our platform? Fill out the form and we'll reach out with next steps.
+                    Want to showcase your products on our platform? Fill out the form and we&apos;ll reach out with next steps.
                   </Text>
                 </Card>
               </Grid.Col>
