+++ app/header/page.tsx
@@ -10,7 +10,7 @@
 
 "use client"
 
-import { AppShell, useMantineTheme, Container, Title, Text, Box } from '@mantine/core';
+import { AppShell, useMantineTheme, Title, Text, Box } from '@mantine/core';
 import NavMT from "@/app/_components/navcomps/navmt";
 import BoxComp from "@/app/_components/othercomps/boxcomp";
 import NewsPage from "@/app/news/page";
