--- app/sign-in/page.tsx
@@ -1,5 +1,0 @@
-// import { SignIn } from "@clerk/nextjs";
-//
-// export default function Page(){
-//     return <SignIn />;
-// }
