+++ app/userfavs/page.tsx
@@ -1,4 +1,4 @@
-import {AppShell, Container, Title} from "@mantine/core";
+import {AppShell, AppShellHeader, AppShellMain, Container, Title} from "@mantine/core";
 import NavMT from "@/app/_components/navcomps/navmt";
 import UserFavs from "@/app/_components/usercomps/userfavs";
 import data from "../_res/usermarkets.json";
@@ -9,13 +9,13 @@ export default async function Page() {
     return (
         <AppShell
         >
-            <AppShell.Header component={NavMT}/>
-            <AppShell.Main>
+            <AppShellHeader component={NavMT}/>
+            <AppShellMain>
                 <Container size="xl" py="xl">
                     <Title order={1} className="text-3xl font-bold mb-6 text-center">{user?.firstName}&apos;s Favourites</Title>
                     <UserFavs data={data}/>
                 </Container>
-            </AppShell.Main>
+            </AppShellMain>
         </AppShell>
     )
 }
